requirements.txt
fastapi
uvicorn
requests
pandas
sqlalchemy
pyodbc
jinja2
python-dotenv